Prepare campaign for C21_WMDE_Test_20
Based on VAR of C21_WMDE_Test_19 ("not now" mini banner).

Copied VAR files to CTRL and deleted VAR components and styles.

Use 1-step form (without address choice) from shared dir and delete all
form-related changes.

Cleaned up old VAR styles (no longer referenced)

https://phabricator.wikimedia.org/T297885
