+++ desktop/banner_ctrl.js
@@ -11,7 +11,7 @@ import { BannerType } from './BannerType';
 import BannerPresenter from '../shared/banner_presenter';
 import Translations from '../shared/messages/de';
 import LocalTranslations from './translations';
-import DonationForm from './components/ui/form/DonationForm';
+import DonationForm from '../shared/components/ui/form/DonationForm';
 import Footer from '../shared/components/ui/EasySelectFooter';
 import BannerText from './components/BannerText';
 import useOfFundsText from '../node_modules/fundraising-frontend-content/i18n/de_DE/data/use_of_funds_content.json';
