+++ desktop/styles/Banner.pcss
@@ -28,6 +28,9 @@
 		font-family: $banner-main-font-family;
 		box-shadow: 0 3px 0.6em rgba( 60, 60, 60, 0.4 );
 		background-color: $color-white;
+		/* Set z-index to hide the animation of mini banner after expanding from mini banner */
+		position: relative;
+		z-index: 200;
 	}
 
 	.banner__content {
@@ -67,16 +70,40 @@
 		text-align: right;
 		line-height: 16px;
 
-		&__link {
-			display: block;
+		&-link {
+			display: inline-block;
+			color: #a9a9a9;
+			font-weight: bold;
+			transition: color 500ms cubic-bezier( 0.63, 0.155, 0.38, 0.885 );
 			width: 100%;
-			font-size: 25px;
-			color: $color-black-light;
 			text-decoration: none;
 			&:hover {
 				text-decoration: none;
 				color: $color-gray-medium;
 			}
+
+			svg {
+				float: right;
+				height: 18px;
+				width: 18px;
+				margin-left: 5px;
+				path {
+					transition: fill 500ms $banner-easing;
+				}
+			}
+
+			&:hover,
+			&:focus {
+				svg path {
+					fill: $color-black;
+				}
+			}
+		}
+	}
+
+	&--content-minimised {
+		.banner__wrapper {
+			display: none;
 		}
 	}
