--- desktop/styles/Banner_var.pcss
@@ -1,116 +1,0 @@
-.wmde-banner {
-	visibility: hidden;
-	width: 100%;
-
-	* {
-		box-sizing: border-box;
-		-webkit-font-smoothing: antialiased;
-		-moz-osx-font-smoothing: grayscale;
-	}
-
-	&--visible {
-		visibility: visible;
-	}
-	&--hidden {
-		display: none;
-	}
-
-	img {
-		border: 0 none;
-	}
-
-	ul {
-		list-style-image: unset;
-	}
-
-	.banner__wrapper {
-		font-size: 14px;
-		font-family: $banner-main-font-family;
-		box-shadow: 0 3px 0.6em rgba( 60, 60, 60, 0.4 );
-		background-color: $color-white;
-		/* Set z-index to hide the animation of mini banner after expanding from mini banner */
-		position: relative;
-		z-index: 200;
-	}
-
-	.banner__content {
-		display: flex;
-		flex-direction: row;
-		order: 1;
-		padding: 0 24px 0;
-	}
-
-	.banner__infobox {
-		display: flex;
-		flex-direction: column;
-		justify-content: space-between;
-		flex: 1 1 auto;
-		margin-bottom: 0;
-		overflow-y: hidden;
-	}
-
-	.banner__form {
-		order: 2;
-		flex: 0 0 $form-width;
-		display: flex;
-		flex-direction: row;
-		width: $form-width;
-		min-height: 273px;
-		.form {
-			width: $form-width;
-		}
-	}
-
-	.banner__footer {
-		order: 2;
-	}
-
-	.close {
-		padding: 8px 8px 0 0;
-		text-align: right;
-		line-height: 16px;
-
-		&-link {
-			display: inline-block;
-			color: #a9a9a9;
-			font-weight: bold;
-			transition: color 500ms cubic-bezier( 0.63, 0.155, 0.38, 0.885 );
-			width: 100%;
-			text-decoration: none;
-			&:hover {
-				text-decoration: none;
-				color: $color-gray-medium;
-			}
-
-			svg {
-				float: right;
-				height: 18px;
-				width: 18px;
-				margin-left: 5px;
-				path {
-					transition: fill 500ms $banner-easing;
-				}
-			}
-
-			&:hover,
-			&:focus {
-				svg path {
-					fill: $color-black;
-				}
-			}
-		}
-	}
-
-	&--content-minimised {
-		.banner__wrapper {
-			display: none;
-		}
-	}
-
-	.banner__slideshow {
-		background: $color-white;
-		display: flex;
-		height: 100%;
-		font-family: $banner-text-font-family;
-	}
-}
