+++ desktop/styles/styles.pcss
@@ -5,15 +5,14 @@
 
 @import './Banner.pcss';
 @import './DonationForm.pcss';
-@import './DonationFormSteps.pcss';
 @import './SelectGroup.pcss';
 @import './Footer.pcss';
 @import './Infobox.pcss';
 @import './BannerText.pcss';
-@import './AnimateTextHighlight.pcss';
 @import './ProgressBar.pcss';
 @import './Slider.pcss';
 @import './AnimateTextHighlight.pcss';
+@import './MinimisedBanner.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsContent.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsDistributionInfo.pcss';
